cpu: rnn: do not mix size_t and int types

+++ src/cpu/rnn/ref_rnn.cpp
@@ -970,7 +970,7 @@ rnn_bias_prepare_sig((_ref_rnn_common_t<aprop, src_type, weights_type,
             scratch_bias_, rnn.n_layer, rnn.n_dir, rnn.n_bias * rnn.dhc);
 
     if (rnn.copy_bias) {
-        parallel_nd(rnn.n_layer * rnn.n_dir, [&](size_t i) {
+        parallel_nd(rnn.n_layer * rnn.n_dir, [&](int i) {
             int off = i * rnn.n_bias * rnn.dhc;
             PRAGMA_OMP_SIMD()
             for (int j = 0; j < rnn.n_bias * rnn.dhc; j++)
